Ton of optimizations for serialization:
Dont include some unset attr like Node.args, kwargs, condition

remove primitives as its own type. Now a list of ints will not have a serialized type for each int. Means during de-serialization we need to check for non nnsight types at the right places.

The intervention object is sent as a string, to be deserialized only when needing it.